refactor: Pull out success key
