+++ backend/api_keys.py
@@ -1,0 +1,2 @@
+MESSAGE_KEY = "message"
+SUCCESS_KEY = "success"
