+++ tests/test_check_stac_metadata_utils_logging.py
@@ -7,7 +7,7 @@ from unittest.mock import MagicMock, patch
 from botocore.exceptions import ClientError  # type: ignore[import]
 from jsonschema import ValidationError  # type: ignore[import]
 
-from backend.api_responses import MESSAGE_KEY
+from backend.api_keys import MESSAGE_KEY, SUCCESS_KEY
 from backend.check_stac_metadata.utils import S3_URL_PREFIX, STACDatasetValidator
 
 from .aws_utils import MockJSONURLReader, MockValidationResultFactory, any_s3_url
@@ -53,7 +53,7 @@ def should_log_non_s3_url_prefix_validation() -> None:
     url_reader = MockJSONURLReader({metadata_url: MINIMAL_VALID_STAC_COLLECTION_OBJECT})
     expected_message = dumps(
         {
-            "success": False,
+            SUCCESS_KEY: False,
             MESSAGE_KEY: f"URL doesn't start with “{S3_URL_PREFIX}”: “{metadata_url}”",
         }
     )
@@ -78,7 +78,7 @@ def should_log_staging_access_validation(validate_mock: MagicMock) -> None:
 
     url_reader = MockJSONURLReader({metadata_url: MINIMAL_VALID_STAC_COLLECTION_OBJECT})
 
-    expected_message = dumps({"success": False, MESSAGE_KEY: str(expected_error)})
+    expected_message = dumps({SUCCESS_KEY: False, MESSAGE_KEY: str(expected_error)})
 
     with patch.object(LOGGER, "error") as logger_mock, patch(
         "backend.check_stac_metadata.utils.processing_assets_model_with_meta"
@@ -98,7 +98,7 @@ def should_log_schema_mismatch_validation(validate_mock: MagicMock) -> None:
 
     url_reader = MockJSONURLReader({metadata_url: MINIMAL_VALID_STAC_COLLECTION_OBJECT})
 
-    expected_message = dumps({"success": False, MESSAGE_KEY: expected_error.message})
+    expected_message = dumps({SUCCESS_KEY: False, MESSAGE_KEY: expected_error.message})
 
     with patch.object(LOGGER, "error") as logger_mock, patch(
         "backend.check_stac_metadata.utils.processing_assets_model_with_meta"
@@ -118,7 +118,7 @@ def should_log_json_parse_validation(validate_mock: MagicMock) -> None:
     expected_error = JSONDecodeError(any_error_message(), "", 0)
     validate_mock.side_effect = expected_error
 
-    expected_message = dumps({"success": False, MESSAGE_KEY: str(expected_error)})
+    expected_message = dumps({SUCCESS_KEY: False, MESSAGE_KEY: str(expected_error)})
 
     with patch.object(LOGGER, "error") as logger_mock, patch(
         "backend.check_stac_metadata.utils.processing_assets_model_with_meta"
