+++ tests/test_import_status_endpoint.py
@@ -7,7 +7,8 @@ from unittest.mock import MagicMock, patch
 
 from pytest import mark
 
-from backend.api_responses import BODY_KEY, HTTP_METHOD_KEY, MESSAGE_KEY, STATUS_CODE_KEY
+from backend.api_keys import MESSAGE_KEY
+from backend.api_responses import BODY_KEY, HTTP_METHOD_KEY, STATUS_CODE_KEY
 from backend.import_file_batch_job_id_keys import ASSET_JOB_ID_KEY, METADATA_JOB_ID_KEY
 from backend.import_status import entrypoint
 from backend.import_status.get import IMPORT_DATASET_KEY, Outcome
